+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [1.26.0] - 2019-04-02
+
 Feature:
 - Add `others` props to `Switch` component
 - Fix: Set `vertical-align` property for `BulletList` component